pyproject.toml
[project]
name = "building-genai-services"
version = "0.1.0"
authors = [
    { name = "David Galera", email = "raizen74@gmail.com" }
]
description = "Building generative ai services with BentoML, FastAPI, and more."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.1",
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "av>=15.1.0",
    "bentoml>=1.4.31",
    "bs4>=0.0.2",
    "diffusers>=0.36.0",
    "fastapi>=0.127.0",
    "greenlet>=3.3.0",
    "python-jose[cryptography]>=3.3.0",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pypdf>=6.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.2",
    "soundfile>=0.13.1",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.50.0",
    "tiktoken>=0.12.0",
    "torch>=2.8.0",
    "transformers>=4.57.3",
    "uvicorn>=0.34.0",
    "bcrypt>=4.0.0,<5.0.0",
]

[project.scripts]
app = "building_genai_services.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
